Initial work with adding Redux
- Ported the auth moduel to use Redux and TypeScript.
- Other related changes.

The exact way we set up our store, ducks and related state will probably
change in future work, so not putting too much effort in it for now.

Using redux-saga for side-effect handling.

The long-running goal is to get rid of Babel and port everything
to using TypeScript. Also replace class components with functional
components using hooks.

Related to #33